requirements.txt
streamlit==1.28.1
langchain==0.1.10
langchain-groq==0.1.1
python-dotenv==1.0.0
